rescale the view's zoom per sequence using v:s
The default behaviour of a view attached to multiple sequences is that
each pixel has the same screen size (or zoom). With v:s, it changes the
current view so that sequences should be displayed with equal size,
which means the screen size of a pixel depends on the size of the images.
This is useful to display images with different resolutions (satellite
bands or multiscale byproducts)