CMakeLists.txt: disable the short hash with MSVC or Ninja
When using Ninja, it seems there is an issue where it will always compute the short hash for each build.

With the Visual Studio generator, it detects cyclic dependencies due to the ZERO_CHECK target.

We need to fix the way the short hash is computed. In the mean time, let's disable this for these generators.